plugins/user.py
from fastapi import APIRouter
from fastapi.responses import JSONResponse
from pyrogram import Client
from pyrogram.enums import ChatType, UserStatus
from pyrogram.errors import PeerIdInvalid, UsernameNotOccupied, ChannelInvalid
from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta
from config import API_ID, API_HASH, BOT_TOKEN
from utils import LOGGER
import asyncio
import threading

router = APIRouter(prefix="/user")
client = None
client_lock = threading.Lock()

def get_dc_locations():
    return {
        1: "MIA, Miami, USA, US",
        2: "AMS, Amsterdam, Netherlands, NL",
        3: "MBA, Mumbai, India, IN",
        4: "STO, Stockholm, Sweden, SE",
        5: "SIN, Singapore, SG",
        6: "LHR, London, United Kingdom, GB",
        7: "FRA, Frankfurt, Germany, DE",
        8: "JFK, New York, USA, US",
        9: "HKG, Hong Kong, HK",
        10: "TYO, Tokyo, Japan, JP",
        11: "SYD, Sydney, Australia, AU",
        12: "GRU, São Paulo, Brazil, BR",
        13: "DXB, Dubai, UAE, AE",
        14: "CDG, Paris, France, FR",
        15: "ICN, Seoul, South Korea, KR",
    }

def calculate_account_age(creation_date):
    today = datetime.now()
    delta = relativedelta(today, creation_date)
    years = delta.years
    months = delta.months
    days = delta.days
    return f"{years} years, {months} months, {days} days"

def estimate_account_creation_date(user_id):
    reference_points = [
        (100000000, datetime(2013, 8, 1)),
        (1273841502, datetime(2020, 8, 13)),
        (1500000000, datetime(2021, 5, 1)),
        (2000000000, datetime(2022, 12, 1)),
    ]
    closest_point = min(reference_points, key=lambda x: abs(x[0] - user_id))
    closest_user_id, closest_date = closest_point
    id_difference = user_id - closest_user_id
    days_difference = id_difference / 20000000
    creation_date = closest_date + timedelta(days=days_difference)
    return creation_date

def format_user_status(status):
    if not status:
        return "Unknown"
    status_map = {
        UserStatus.ONLINE: "Online",
        UserStatus.OFFLINE: "Offline",
        UserStatus.RECENTLY: "Recently online",
        UserStatus.LAST_WEEK: "Last seen within week",
        UserStatus.LAST_MONTH: "Last seen within month"
    }
    return status_map.get(status, "Unknown")

async def ensure_client():
    global client
    with client_lock:
        if client is None:
            try:
                client = Client(
                    "A360APIUser",
                    api_id=API_ID,
                    api_hash=API_HASH,
                    bot_token=BOT_TOKEN
                )
                await client.start()
                LOGGER.info("Pyrogram client started successfully")
                return True
            except Exception as e:
                LOGGER.error(f"Failed to start Pyrogram client: {str(e)}")
                client = None
                return False
        
        try:
            is_connected = getattr(client, 'is_connected', False)
            if callable(is_connected):
                connected = is_connected()
            else:
                connected = is_connected
            
            if not connected:
                await client.start()
                LOGGER.info("Pyrogram client reconnected successfully")
        except Exception as e:
            LOGGER.error(f"Failed to check/restart client: {str(e)}")
            try:
                client = Client(
                    "A360APIUser",
                    api_id=API_ID,
                    api_hash=API_HASH,
                    bot_token=BOT_TOKEN
                )
                await client.start()
                LOGGER.info("Pyrogram client recreated successfully")
            except Exception as e2:
                LOGGER.error(f"Failed to recreate client: {str(e2)}")
                client = None
                return False
        
        return True

async def get_user_info(username):
    try:
        if not await ensure_client():
            return {"success": False, "error": "Client initialization failed"}
        
        DC_LOCATIONS = get_dc_locations()
        user = await client.get_users(username)
        premium_status = getattr(user, 'is_premium', False)
        dc_location = DC_LOCATIONS.get(user.dc_id, "Unknown")
        account_created = estimate_account_creation_date(user.id)
        account_created_str = account_created.strftime("%B %d, %Y")
        account_age = calculate_account_age(account_created)
        verified_status = getattr(user, 'is_verified', False)
        status = format_user_status(getattr(user, 'status', None))
        flags = "Clean"
        if getattr(user, 'is_scam', False):
            flags = "Scam"
        elif getattr(user, 'is_fake', False):
            flags = "Fake"
        
        user_data = {
            "success": True,
            "type": "bot" if user.is_bot else "user",
            "id": user.id,
            "first_name": user.first_name,
            "last_name": user.last_name,
            "username": user.username,
            "dc_id": user.dc_id,
            "dc_location": dc_location,
            "is_premium": premium_status,
            "is_verified": verified_status,
            "is_bot": user.is_bot,
            "flags": flags,
            "status": status,
            "account_created": account_created_str,
            "account_age": account_age,
            "api_owner": "@ISmartCoder",
            "api_updates": "t.me/TheSmartDev",
            "links": {
                "android": f"tg://openmessage?user_id={user.id}",
                "ios": f"tg://user?id={user.id}",
                "permanent": f"tg://user?id={user.id}"
            }
        }
        return user_data
    except (PeerIdInvalid, UsernameNotOccupied):
        return {"success": False, "error": "User not found"}
    except Exception as e:
        LOGGER.error(f"Error fetching user info: {str(e)}")
        return {"success": False, "error": f"Failed to fetch user information: {str(e)}"}

async def get_chat_info(username):
    try:
        if not await ensure_client():
            return {"success": False, "error": "Client initialization failed"}
        
        DC_LOCATIONS = get_dc_locations()
        chat = await client.get_chat(username)
        chat_type_map = {
            ChatType.SUPERGROUP: "supergroup",
            ChatType.GROUP: "group",
            ChatType.CHANNEL: "channel"
        }
        chat_type = chat_type_map.get(chat.type, "unknown")
        dc_location = DC_LOCATIONS.get(getattr(chat, 'dc_id', None), "Unknown")
        
        if chat.username:
            join_link = f"t.me/{chat.username}"
            permanent_link = f"t.me/{chat.username}"
        elif chat.id < 0:
            chat_id_str = str(chat.id).replace('-100', '')
            join_link = f"t.me/c/{chat_id_str}/1"
            permanent_link = f"t.me/c/{chat_id_str}/1"
        else:
            join_link = f"tg://resolve?domain={chat.id}"
            permanent_link = f"tg://resolve?domain={chat.id}"
        
        chat_data = {
            "success": True,
            "type": chat_type,
            "id": chat.id,
            "title": chat.title,
            "username": chat.username,
            "dc_id": getattr(chat, 'dc_id', None),
            "dc_location": dc_location,
            "members_count": getattr(chat, 'members_count', None),
            "description": getattr(chat, 'description', None),
            "api_owner": "@ISmartCoder",
            "api_updates": "t.me/TheSmartDev",
            "links": {
                "join": join_link,
                "permanent": permanent_link
            }
        }
        return chat_data
    except (ChannelInvalid, PeerIdInvalid):
        return {"success": False, "error": "Chat not found or access denied"}
    except Exception as e:
        LOGGER.error(f"Error fetching chat info: {str(e)}")
        return {"success": False, "error": f"Failed to fetch chat information: {str(e)}"}

async def get_telegram_info(username):
    username = username.strip('@').replace('https://', '').replace('http://', '').replace('t.me/', '').replace('/', '').replace(':', '')
    LOGGER.info(f"Fetching info for: {username}")
    
    user_info = await get_user_info(username)
    if user_info["success"]:
        return user_info
    
    chat_info = await get_chat_info(username)
    if chat_info["success"]:
        return chat_info
    
    return {"success": False, "error": "Entity not found in Telegram database", "api_owner": "@ISmartCoder", "api_updates": "t.me/TheSmartDev"}

@router.get("/info")
async def info_endpoint(username: str = ""):
    if not username:
        return JSONResponse(
            status_code=400,
            content={
                "success": False,
                "error": "Missing 'username' parameter",
                "api_owner": "@ISmartCoder",
                "api_updates": "t.me/TheSmartDev"
            }
        )
    try:
        result = await get_telegram_info(username)
        return JSONResponse(
            content=result,
            status_code=200 if result["success"] else 404
        )
    except Exception as e:
        LOGGER.error(f"Unexpected error fetching Telegram info for {username}: {str(e)}")
        return JSONResponse(
            status_code=500,
            content={
                "success": False,
                "error": f"Internal server error: {str(e)}",
                "api_owner": "@ISmartCoder",
                "api_updates": "t.me/TheSmartDev"
            }
        )