requirements.txt
pyrofork
tgcyrpto
telethon 
cyrptg
fastapi
uvicorn
asyncio
smartfaker
smartbindb
googletrans==4.0.0-rc1
beautifulsoup4
cloudscraper
requests 
aiohttp
pycountry 
pydantic
python-dateutil
urllib